requirements.txt
python-dotenv==0.14.0
requests==2.24.0
instabot==0.117.0